pyproject.toml
[tool.poetry]
name = "vigilant-pancake"
version = "0.1.0"
description = ""
authors = ["Kostiantyn Ivashchenko <kivaschenko@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^8.3.1"
requests = "^2.32.3"
SQLAlchemy = "^2.0.31"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
fastapi = "^0.111.1"
uvicorn = "^0.30.3"
pydantic-settings = "^2.3.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
